flask-webserver/pyproject.toml
[project]
name = "flask-webserver"
version = "0.1.0"
description = "Flask + Resonate example"
readme = "README.md"
requires-python = ">= 3.12"
authors = [
    { name = "ResonateHQ", email = "contact@resonatehq.io" }
]
dependencies = [
    "flask>=3.0.3",
    "resonate-sdk>=0.4.12",
]

[project.scripts]
"webserver" = "src.webserver:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
